Add files via upload
Analyzing historical stock revenue data and building a dashboard

Extracting essential data from a dataset and displaying it is a necessary part of data science; therefore individuals can make correct decisions based on the data. In this assignment, we will extract some stock data, we will then display this data in a graph.